harness/pyproject.toml
[project]
name = "dag-harness"
version = "0.3.0"
description = "Disposable DAG orchestration harness with LangGraph and MCP client integration"
readme = "../README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jess Sullivan", email = "jess@tinyland.dev"}
]
maintainers = [
    {name = "Tinyland.dev, Inc.", email = "hello@tinyland.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
keywords = ["ansible", "dag", "langgraph", "mcp-client", "orchestration"]

dependencies = [
    # LangGraph 1.0.x - stable release with RetryPolicy support
    # Migrated from 0.2.x in Task #13
    # Breaking changes addressed:
    # - RetryPolicy import from langgraph.pregel (new in 1.0.x)
    # - AsyncSqliteSaver path unchanged (langgraph.checkpoint.sqlite.aio)
    # - StateGraph, END imports unchanged
    "langgraph>=1.0.7",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain-core>=0.3.0",
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0.0",
    "packaging>=23.0",
]

[project.urls]
Homepage = "https://disposable-ansible-dag.ephemera.xoxd.ai"
Documentation = "https://disposable-ansible-dag.ephemera.xoxd.ai"
Repository = "https://github.com/Jesssullivan/dag-harness"
GitLab = "https://gitlab.com/tinyland/projects/dag-harness"
Issues = "https://github.com/Jesssullivan/dag-harness/issues"
Changelog = "https://github.com/Jesssullivan/dag-harness/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "ruff>=0.4.0",
]
proxy = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]
notifications = [
    "aiosmtplib>=3.0.0",
]
claude-sdk = [
    # Claude Agent SDK for native integration with Claude Code
    # Replaces subprocess spawning with in-process API
    "claude-agent-sdk>=0.1.0",
]
postgres = [
    # PostgreSQL checkpointer for LangGraph (production-ready)
    # Provides async connection pooling and checkpoint persistence
    "langgraph-checkpoint-postgres>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary,pool]>=3.1.0",
]
observability = [
    # LangSmith for tracing and observability
    # Provides automatic trace collection when LANGCHAIN_TRACING_V2=true
    "langsmith>=0.1.0",
]
sandbox = [
    # E2B sandbox for isolated code execution
    # Provides secure execution environment for destructive operations
    "e2b>=0.17.0",
    "e2b-code-interpreter>=0.0.11",
]
all = [
    "dag-harness[dev,proxy,notifications,claude-sdk,observability,postgres,sandbox]",
]

[project.scripts]
harness = "harness.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "harness/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["harness"]

[tool.hatch.build.targets.wheel.force-include]
# Include SQL schemas for database initialization
"harness/db/schema.sql" = "harness/db/schema.sql"
# Include templates for harness init
"harness/templates/harness.yml.j2" = "harness/templates/harness.yml.j2"
"harness/templates/gitignore_additions.txt" = "harness/templates/gitignore_additions.txt"

[tool.hatch.build.targets.sdist]
include = [
    "/harness",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic",
    "ignore::DeprecationWarning:langchain_core",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",   # Line too long
    "E402",   # Module level import not at top (needed for type aliases)
    "N806",   # Variable in function should be lowercase (test conventions)
    "N812",   # Lowercase imported as non-lowercase (version constants)
    "N813",   # Camelcase imported as lowercase (test imports)
]

[tool.ruff.lint.per-file-ignores]
# Bootstrap module uses imports for availability checking
"harness/bootstrap/__init__.py" = ["F401"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
changelog_file = "../CHANGELOG.md"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variable = ["harness/__init__.py:__version__"]
branch = "main"
build_command = "uv build"
commit_message = "chore(release): bump version to {version}"

[tool.semantic_release.changelog]
changelog_file = "../CHANGELOG.md"

[tool.semantic_release.remote]
type = "github"

[tool.coverage.run]
source = ["harness"]
omit = ["harness/__pycache__/*", "tests/*"]
